+++ Controller/ResettingController.php
@@ -11,6 +11,7 @@
 
 namespace FOS\UserBundle\Controller;
 
+use FOS\UserBundle\CompatibilityUtil;
 use FOS\UserBundle\Event\FilterUserResponseEvent;
 use FOS\UserBundle\Event\FormEvent;
 use FOS\UserBundle\Event\GetResponseNullableUserEvent;
@@ -21,11 +22,10 @@ use FOS\UserBundle\Mailer\MailerInterface;
 use FOS\UserBundle\Model\UserManagerInterface;
 use FOS\UserBundle\Util\TokenGeneratorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Symfony\Component\EventDispatcher\LegacyEventDispatcherProxy;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
 /**
  * Controller managing the resetting of the password.
@@ -53,7 +53,7 @@ class ResettingController extends AbstractController
      */
     public function __construct(EventDispatcherInterface $eventDispatcher, FactoryInterface $formFactory, UserManagerInterface $userManager, TokenGeneratorInterface $tokenGenerator, MailerInterface $mailer, $retryTtl)
     {
-        $this->eventDispatcher = LegacyEventDispatcherProxy::decorate($eventDispatcher);
+        $this->eventDispatcher = CompatibilityUtil::upgradeEventDispatcher($eventDispatcher);
         $this->formFactory = $formFactory;
         $this->userManager = $userManager;
         $this->tokenGenerator = $tokenGenerator;
@@ -64,17 +64,15 @@ class ResettingController extends AbstractController
     /**
      * Request reset user password: show form.
      */
-    public function requestAction()
+    public function requestAction(): Response
     {
         return $this->render('@FOSUser/Resetting/request.html.twig');
     }
 
     /**
      * Request reset user password: submit form and send email.
-     *
-     * @return Response
      */
-    public function sendEmailAction(Request $request)
+    public function sendEmailAction(Request $request): Response
     {
         $username = $request->request->get('username');
 
@@ -123,10 +121,8 @@ class ResettingController extends AbstractController
 
     /**
      * Tell the user to check his email provider.
-     *
-     * @return Response
      */
-    public function checkEmailAction(Request $request)
+    public function checkEmailAction(Request $request): Response
     {
         $username = $request->query->get('username');
 
@@ -144,10 +140,8 @@ class ResettingController extends AbstractController
      * Reset user password.
      *
      * @param string $token
-     *
-     * @return Response
      */
-    public function resetAction(Request $request, $token)
+    public function resetAction(Request $request, $token): Response
     {
         $user = $this->userManager->findUserByConfirmationToken($token);