+++ Event/GetResponseNullableUserEvent.php
@@ -12,29 +12,56 @@
 namespace FOS\UserBundle\Event;
 
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Contracts\EventDispatcher\Event;
 
 /**
  * Response user event that allows null user.
  *
  * @author Konstantinos Christofilos <kostas.christofilos@gmail.com>
+ * @final
  */
-class GetResponseNullableUserEvent extends GetResponseUserEvent
+class GetResponseNullableUserEvent extends Event
 {
     /**
-     * GetResponseNullableUserEvent constructor.
+     * @var Request
      */
-    public function __construct(UserInterface $user = null, Request $request)
+    private $request;
+
+    /**
+     * @var UserInterface|null
+     */
+    private $user;
+
+    /**
+     * @var Response|null
+     */
+    private $response;
+
+    public function __construct(?UserInterface $user, Request $request)
     {
         $this->user = $user;
         $this->request = $request;
     }
 
-    /**
-     * @return UserInterface|null
-     */
-    public function getUser()
+    public function getUser(): ?UserInterface
     {
         return $this->user;
+    }
+
+    public function getRequest(): Request
+    {
+        return $this->request;
+    }
+
+    public function setResponse(Response $response)
+    {
+        $this->response = $response;
+    }
+
+    public function getResponse(): ?Response
+    {
+        return $this->response;
     }
 }