+++ EventListener/FlashListener.php
@@ -12,11 +12,11 @@
 namespace FOS\UserBundle\EventListener;
 
 use FOS\UserBundle\FOSUserEvents;
-use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\HttpFoundation\Session\SessionInterface;
-use Symfony\Component\Translation\TranslatorInterface;
-use Symfony\Component\Translation\DataCollectorTranslator;
+use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\HttpFoundation\Session\Session;
+use Symfony\Contracts\EventDispatcher\Event;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * @internal
@@ -38,9 +38,9 @@ class FlashListener implements EventSubscriberInterface
     ];
 
     /**
-     * @var SessionInterface
+     * @var RequestStack
      */
-    private $session;
+    private $requestStack;
 
     /**
      * @var TranslatorInterface
@@ -50,16 +50,16 @@ class FlashListener implements EventSubscriberInterface
     /**
      * FlashListener constructor.
      */
-    public function __construct(SessionInterface $session, DataCollectorTranslator $translator)
+    public function __construct(RequestStack $requestStack, TranslatorInterface $translator)
     {
-        $this->session = $session;
         $this->translator = $translator;
+        $this->requestStack = $requestStack;
     }
 
     /**
      * {@inheritdoc}
      */
-    public static function getSubscribedEvents()
+    public static function getSubscribedEvents(): array
     {
         return [
             FOSUserEvents::CHANGE_PASSWORD_COMPLETED => 'addSuccessFlash',
@@ -81,15 +81,24 @@ class FlashListener implements EventSubscriberInterface
             throw new \InvalidArgumentException('This event does not correspond to a known flash message');
         }
 
-        $this->session->getFlashBag()->add('success', $this->trans(self::$successMessages[$eventName]));
+        $this->getSession()->getFlashBag()->add('success', $this->trans(self::$successMessages[$eventName]));
+    }
+
+    private function getSession(): Session
+    {
+        $request = $this->requestStack->getCurrentRequest();
+
+        if (null === $request) {
+            throw new \LogicException('Cannot get the session without an active request.');
+        }
+
+        return $request->getSession();
     }
 
     /**
-     * @param string$message
-     *
-     * @return string
+     * @param string $message
      */
-    private function trans($message, array $params = [])
+    private function trans($message, array $params = []): string
     {
         return $this->translator->trans($message, $params, 'FOSUserBundle');
     }